+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/ClientModInitializerImpl.java
@@ -45,7 +45,6 @@ import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.fabricmc.fabric.api.client.rendering.v1.TooltipComponentCallback;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.client.KeyMapping;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.MenuScreens;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.MenuAccess;
@@ -54,7 +53,6 @@ import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderers;
 import net.minecraft.client.renderer.item.ItemProperties;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.DyeColor;
@@ -296,20 +294,7 @@ public class ClientModInitializerImpl extends AbstractClientModInitializer imple
             InputConstants.UNKNOWN.getValue(),
             KEY_BINDINGS_TRANSLATION_KEY
         )));
-        ClientTickEvents.END_CLIENT_TICK.register(client -> {
-            final Player player = Minecraft.getInstance().player;
-            if (player == null) {
-                return;
-            }
-            final KeyMapping openWirelessGrid = KeyMappings.INSTANCE.getOpenWirelessGrid();
-            while (openWirelessGrid != null && openWirelessGrid.consumeClick()) {
-                PlatformApi.INSTANCE.useNetworkBoundItem(
-                    player,
-                    Items.INSTANCE.getWirelessGrid(),
-                    Items.INSTANCE.getCreativeWirelessGrid()
-                );
-            }
-        });
+        ClientTickEvents.END_CLIENT_TICK.register(client -> handleInputEvents());
     }
 
     private void registerModelPredicates() {